Refactor proc macro code
